fix: ensure that program works for different precisions

+++ src/types.hpp
@@ -1,4 +1,4 @@
 #pragma once
 
 // using Decimal = float;
-using Decimal = double;
+using Decimal = long double;
